refactor: 로또 번호 검증에서 유저 로또 번호 입력 검증을 분리
