+++ src/main/java/lotto/view/InputView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
-import lotto.domain.validation.LottoNumbersValidation;
+import lotto.view.validation.InputNumbersValidation;
 
 public class InputView {
     public static String buyLotto() {
@@ -13,7 +13,7 @@ public class InputView {
     public static List<Integer> inputLottoNumbers() {
         System.out.println(GameMessage.INPUT_NUMBERS);
         String userNumbers = Console.readLine();
-        LottoNumbersValidation.validateInputLottoNumbers(userNumbers);
+        InputNumbersValidation .validate(userNumbers);
         return InputUtil.convertUserInputToNumbers(userNumbers);
     }
